Add data type hints to column and entry descriptions
Fixes #35

With this, it is expected that column and entry descriptions may contain
an optional typing option, to explain what the data type is and, if
available, what unit to use when formatting it.

This allow the server implementations to send raw data to the client,
who would then have the option to format the data, but also the
possibility to do some other actions, like sorting, or use the data for
richer data interactions, as raw data may have a same meaning in
different outputs, unlike pre-formatted data which may not allow such
interaction.

Signed-off-by: Geneviève Bastien <gbastien@versatic.net>