Add a swagger/index for server's openapi.yaml TSP
This change is the companion of the master branch commit that introduces
openapi.yaml as the alternate TSP version. This is the TSP documented by
Swagger now supported directly in trace-server. As explained in that
master commit's message, these alternate GH pages for TSP show the true
endpoints currently implemented in server. There is that plan to further
align the former manual API.yaml with this generated openapi.yaml,
step-wise.

The master branch with that companion commit is to be merged in this
branch, as usual. That merge should bring the companion openapi.yaml
file and README instructions along.

Fixes #61 partially as first steps.

Signed-off-by: Marco Miller <marco.miller@ericsson.com>