openapi: Remove leftover QueryParameters few lines
Have openapi.py remove the only few left-over QueryParameters lines from
openapi.yaml. Remove these lines using this python script, per README.md
instructions, as companion [1] (swagger-core, server) cannot remove them
all. Include an updated openapi.yaml based on this and [1] altogether.

Should the python script find Filter component lines present in
openapi.yaml, this means that the QueryParameters component was not
hidden by swagger-core (in server). -QueryParameters being the one that
brings Filter in. In this unexpected case of a misbehaving or unpatched
server, without [1] that is, do not proceed with the hereby removal. The
goal of this being, to not partially remove component lines thus corrupt
the resulting yaml.

[1] https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/189088

Signed-off-by: Marco Miller <marco.miller@ericsson.com>