Introduce ErrorResponse and use it everywhere for error responses
When an error is created when processing a client request, return
an ErrorResponse JSON object that includes the detailed error message.

This standardizes the error response to common data structure.

This commit also updates the swagger documentation accordingly.

This commit also fixes the swagger documentation for trace and
experiment regarding error code 209 (conflict) and some missing
definitions.

Fixes #122:

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>